Added saveCycleData and convert singles to double on k2dsm

+++ +Utils/+convert/applyAcResOnDsmModel.m
@@ -6,15 +6,15 @@ function dsmRegsOut = applyAcResOnDsmModel(acData, dsmRegsIn, type)
                 case 0 % none
                     dsmRegsOut = dsmRegsIn;
                 case 1 % AOT model
-                    dsmRegsOut.dsmXscale = dsmRegsIn.dsmXscale * acData.hFactor;
-                    dsmRegsOut.dsmYscale = dsmRegsIn.dsmYscale * acData.vFactor;
-                    dsmRegsOut.dsmXoffset = (dsmRegsIn.dsmXoffset + acData.hOffset) / acData.hFactor;
-                    dsmRegsOut.dsmYoffset = (dsmRegsIn.dsmYoffset + acData.vOffset) / acData.vFactor;
+                    dsmRegsOut.dsmXscale = double(dsmRegsIn.dsmXscale) * double(acData.hFactor);
+                    dsmRegsOut.dsmYscale = double(dsmRegsIn.dsmYscale) * double(acData.vFactor);
+                    dsmRegsOut.dsmXoffset = (double(dsmRegsIn.dsmXoffset) + double(acData.hOffset)) / double(acData.hFactor);
+                    dsmRegsOut.dsmYoffset = (double(dsmRegsIn.dsmYoffset) + double(acData.vOffset)) / double(acData.vFactor);
                 case 2 % TOA model
-                    dsmRegsOut.dsmXscale = dsmRegsIn.dsmXscale * acData.hFactor;
-                    dsmRegsOut.dsmYscale = dsmRegsIn.dsmYscale * acData.vFactor;
-                    dsmRegsOut.dsmXoffset = dsmRegsIn.dsmXoffset + acData.hOffset / dsmRegsIn.dsmXscale;
-                    dsmRegsOut.dsmYoffset = dsmRegsIn.dsmYoffset + acData.vOffset / dsmRegsIn.dsmYscale;
+                    dsmRegsOut.dsmXscale = double(dsmRegsIn.dsmXscale) * double(acData.hFactor);
+                    dsmRegsOut.dsmYscale = double(dsmRegsIn.dsmYscale) * double(acData.vFactor);
+                    dsmRegsOut.dsmXoffset = double(dsmRegsIn.dsmXoffset) + double(acData.hOffset) / double(dsmRegsIn.dsmXscale);
+                    dsmRegsOut.dsmYoffset = double(dsmRegsIn.dsmYoffset) + double(acData.vOffset) / double(dsmRegsIn.dsmYscale);
                 otherwise
                     error('Only {0,1,2} are supported as values for "flags"');
             end
@@ -23,15 +23,15 @@ function dsmRegsOut = applyAcResOnDsmModel(acData, dsmRegsIn, type)
                 case 0 % none
                     dsmRegsOut = dsmRegsIn;
                 case 1 % AOT model
-                    dsmRegsOut.dsmXscale = dsmRegsIn.dsmXscale / acData.hFactor;
-                    dsmRegsOut.dsmYscale = dsmRegsIn.dsmYscale / acData.vFactor;
-                    dsmRegsOut.dsmXoffset = dsmRegsIn.dsmXoffset * acData.hFactor - acData.hOffset;
-                    dsmRegsOut.dsmYoffset = dsmRegsIn.dsmYoffset * acData.vFactor - acData.vOffset;
+                    dsmRegsOut.dsmXscale = double(dsmRegsIn.dsmXscale) / double(acData.hFactor);
+                    dsmRegsOut.dsmYscale = double(dsmRegsIn.dsmYscale) / double(acData.vFactor);
+                    dsmRegsOut.dsmXoffset = double(dsmRegsIn.dsmXoffset) * double(acData.hFactor) - double(acData.hOffset);
+                    dsmRegsOut.dsmYoffset = double(dsmRegsIn.dsmYoffset) * double(acData.vFactor) - double(acData.vOffset);
                 case 2 % TOA model
-                    dsmRegsOut.dsmXscale = dsmRegsIn.dsmXscale / acData.hFactor;
-                    dsmRegsOut.dsmYscale = dsmRegsIn.dsmYscale / acData.vFactor;
-                    dsmRegsOut.dsmXoffset = dsmRegsIn.dsmXoffset - acData.hOffset / dsmRegsOut.dsmXscale;
-                    dsmRegsOut.dsmYoffset = dsmRegsIn.dsmYoffset - acData.vOffset / dsmRegsOut.dsmYscale;
+                    dsmRegsOut.dsmXscale = double(dsmRegsIn.dsmXscale) / double(acData.hFactor);
+                    dsmRegsOut.dsmYscale = double(dsmRegsIn.dsmYscale) / double(acData.vFactor);
+                    dsmRegsOut.dsmXoffset = double(dsmRegsIn.dsmXoffset) - double(acData.hOffset) / double(dsmRegsOut.dsmXscale);
+                    dsmRegsOut.dsmYoffset = double(dsmRegsIn.dsmYoffset) - double(acData.vOffset) / double(dsmRegsOut.dsmYscale);
                 otherwise
                     error('Only {0,1,2} are supported as values for "flags"');
             end
